requirements.txt
Flask
Flask-Bootstrap
Flask-Migrate
Flask-SQLAlchemy
Jinja2
numpy
scipy
SQLAlchemy
opencv-python
pillow